+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Added `/stats`, `/rich`, `/guildconfig` and `/say` commands.
 + Added multiple sources to `/lore` command.
 + `/handbook` available as a context-menu command on users.
++ Basic I18n is introduced. Language can now be set on a per-guild basis.
 
 ## 0.2.1